+++ include/afio/revision.hpp
@@ -1,4 +1,4 @@
 // Note the second line of this file must ALWAYS be the git SHA, third line ALWAYS the git SHA update time
-#define AFIO_PREVIOUS_COMMIT_REF    dfe971cb67c8af1d86059552aef3c9420c915dfb
-#define AFIO_PREVIOUS_COMMIT_DATE   "2017-08-09 14:09:12 +00:00"
-#define AFIO_PREVIOUS_COMMIT_UNIQUE dfe971cb
+#define AFIO_PREVIOUS_COMMIT_REF    9d82063bf3cb04966cf7d129e4e50a75779e90b0
+#define AFIO_PREVIOUS_COMMIT_DATE   "2017-08-10 22:42:56 +00:00"
+#define AFIO_PREVIOUS_COMMIT_UNIQUE 9d82063b